Move files to valide folders
