+++ classes/music_album.rb
@@ -1,5 +1,5 @@
-require './item'
 require 'json'
+
 class MusicAlbum < Item
   attr_accessor :on_spotify
