+++ controllers/genre_controller.rb
@@ -1,4 +1,4 @@
-require_relative '../genre'
+
 class GenreController
   attr_reader :genres
 
@@ -11,7 +11,7 @@ class GenreController
       puts ' Genre list is empty! choose the option to add a genre from the list'
     else
       @genres.each_with_index do |genre, index|
-        puts "- Genre#{index + 1} id: #{genre.id} name: #{genre.name}"
+        puts "- index #{index + 1} id: #{genre.id} name: #{genre.name}"
       end
     end
     @genres.length
