--- schema/book_schema.sql
@@ -1,15 +1,0 @@
-CREATE TABLE books (
-  id INT PRIMARY KEY NOT NULL GENERATED ALWAYS AS IDENTITY,
-  author_id INT,
-  genre_id INT,
-  source_id INT,
-  label_id INT,
-	FOREIGN KEY (genre_id) REFERENCES genre (id) ON DELETE SET NULL,
-	FOREIGN KEY (author_id) REFERENCES author (id) ON DELETE SET NULL,
-	FOREIGN KEY (source_id) REFERENCES source (id) ON DELETE SET NULL,
-  FOREIGN KEY (label_id) REFERENCES labels (id) ON DELETE SET NULL,
-  publish_date DATE,
-  archived BOOLEAN,
-  publisher VARCHAR(200),
-  cover_state VARCHAR(200),
-);
