--- schema/label_schema.sql
@@ -1,7 +1,0 @@
-CREATE TABLE labels (
-  id INT PRIMARY KEY NOT NULL GENERATED ALWAYS AS IDENTITY,
-  title VARCHAR(200),
-  color VARCHAR(200),
-  item_id INT,
-	FOREIGN KEY (item_id) REFERENCES item (id) ON DELETE SET NULL,
-);
