+++ storage/game.json
@@ -1,1 +1,1 @@
-["{\"id\":730,\"multiplayer\":null,\"last_played_at\":\"2020-05-03\",\"publish_date\":\"2010-05-02\"}"]
+["{\"id\":730,\"multiplayer\":null,\"last_played_at\":\"2020-05-03\",\"publish_date\":\"2010-05-02\"}","{\"id\":727,\"multiplayer\":null,\"last_played_at\":\"2022-06-06\",\"publish_date\":\"2019-05-02\"}"]
