+++ storage/labels.json
@@ -1,1 +1,1 @@
-[{"json_class":"Label","a":["ll","aa",1462]},{"json_class":"Label","a":["ALL","red",1527]}]
+[{"json_class":"Label","a":["ll","aa",1532]},{"json_class":"Label","a":["ALL","red",306]},{"json_class":"Label","a":["lan","green",169]},{"json_class":"Label","a":["AA","yellow",368]}]
