+++ storage/music_album.json
@@ -1,1 +1,1 @@
-["{\"id\":972,\"publish_date\":\"2021-05-20\",\"on_spotify\":true,\"genre\":null,\"archived\":false}","{\"id\":75,\"publish_date\":\"2019-03-20\",\"on_spotify\":false,\"genre\":null,\"archived\":false}","{\"id\":981,\"publish_date\":\"2010-04-30\",\"on_spotify\":true,\"genre\":null,\"archived\":false}","{\"id\":969,\"publish_date\":\"2021-06-13\",\"on_spotify\":true,\"genre\":null,\"archived\":false}","{\"id\":618,\"publish_date\":\"2010-06-22\",\"on_spotify\":true,\"genre\":null,\"archived\":false}"]
+["{\"id\":908,\"publish_date\":\"2021-05-20\",\"on_spotify\":true,\"genre\":null,\"archived\":false}","{\"id\":204,\"publish_date\":\"2019-03-20\",\"on_spotify\":false,\"genre\":null,\"archived\":false}","{\"id\":102,\"publish_date\":\"2010-04-30\",\"on_spotify\":true,\"genre\":null,\"archived\":false}","{\"id\":958,\"publish_date\":\"2021-06-13\",\"on_spotify\":true,\"genre\":null,\"archived\":false}","{\"id\":79,\"publish_date\":\"2010-06-22\",\"on_spotify\":true,\"genre\":null,\"archived\":false}","{\"id\":708,\"publish_date\":\"2010-05-03\",\"on_spotify\":true,\"genre\":null,\"archived\":false}","{\"id\":889,\"publish_date\":\"2020-05-03\",\"on_spotify\":true,\"genre\":null,\"archived\":false}"]
